server: ruff bandit/SIM/N/RUF/PTH rule sets, housekeeping (T3, T4, D7)
T3: enabled S (bandit), SIM, N, RUF, and PTH ruff rule sets. Fixed the
29 real findings (pathlib over os.path, itertools.pairwise, collapsed
nested ifs, sorted __all__, an already-int cast, SQLAlchemy test
variable naming) and noqa'd the handful of genuine false positives
(0.0.0.0 bind in a container entrypoint, a token label prefix that
isn't the actual secret, an assert enforcing an argparse invariant,
seeded non-crypto random in a deterministic test fixture). Renamed
three exceptions to satisfy N818's Error-suffix convention
(SecretKeyNotConfiguredError, InvalidCursorError, ValidationFailedError)
and removed the unused token_hashes_equal helper. Extracted the
repeated Alembic-migration-against-a-tmp-DB boilerplate (three
near-identical copies) into tests/conftest.py's run_migrations().

T4: deleted the stray empty android/ directory (gitignored leftover
from the W0 mobile/ move); added LICENSE (MIT) and SECURITY.md.

D7: expanded the "Using a different reverse proxy" README section with
the concrete contract any proxy must satisfy (required headers,
SR_TRUSTED_PROXIES precision, body size, no self-redirect, no
websocket need) and a worked Caddyfile example.

+++ server/app/web/formatting.py
@@ -18,7 +18,7 @@ def format_distance_km(meters: float | None) -> str:
 def format_duration(seconds: float | None) -> str:
     if seconds is None or seconds < 0:
         return "—"
-    total = int(round(seconds))
+    total = round(seconds)
     hours, remainder = divmod(total, 3600)
     minutes, secs = divmod(remainder, 60)
     if hours:
@@ -31,5 +31,5 @@ def format_pace(speed_mps: float | None) -> str:
     if speed_mps is None or speed_mps <= 0:
         return "—"
     seconds_per_km = 1000 / speed_mps
-    minutes, secs = divmod(int(round(seconds_per_km)), 60)
+    minutes, secs = divmod(round(seconds_per_km), 60)
     return f"{minutes}:{secs:02d} /km"
